Uninstalled styled-components - some visual bug fixes
