+++ web/package.json
@@ -39,8 +39,6 @@
     "react-modal-sheet": "1.8.1",
     "react-player": "^2.11.0",
     "react-query": "^3.39.2",
-    "styled-components": "^5.3.6",
-    "styled-normalize": "^8.0.7",
     "superjson": "^1.12.1",
     "uifx": "^2.0.7",
     "zod": "^3.20.2",
@@ -54,7 +52,6 @@
     "@types/prismjs": "^1.16.5",
     "@types/react": "^18.0.17",
     "@types/react-dom": "^18.0.6",
-    "@types/styled-components": "^5.1.7",
     "autoprefixer": "^10.4.13",
     "csv-parse": "^5.3.3",
     "eslint": "^8.31.0",
