+++ web/src/components/Players/AudioPlayer.tsx
@@ -19,8 +19,8 @@ export const AudioPlayer = (
         next: <Icon icon="next" />,
         loop: <Icon icon="loop" />,
         loopOff: <Icon icon="loopOff" />,
-        volume: <Icon icon="volume" size={20} />,
-        volumeMute: <Icon icon="volumeMute" size={20} />,
+        volume: <Icon icon="volume" className="child:w-5 child:h-5" size="class" />,
+        volumeMute: <Icon icon="volumeMute" className="child:w-5 child:h-5" size="class" />,
       }}
       {...props}
     />
