+++ web/src/lib/game/steps/Step2.tsx
@@ -3,8 +3,7 @@ import Image from "next/image";
 import { classed } from "@tw-classed/react"
 import sample from "lodash/sample";
 import Sheet from "react-modal-sheet";
-import { theme, useIsDesktop } from "~/styles/theme";
-import { bounceIn, bounceOut, fadeUpIn, drawIn } from "~/styles/animations";
+import { useIsDesktop } from "~/utils/useIsDesktop";
 import { Instruction } from "~/lib/game/components/Instruction";
 import { Question } from "~/lib/game/components/Question";
 import { Answer } from "~/lib/game/components/Answer";
@@ -14,7 +13,7 @@ import {
   NextButton,
 } from "~/lib/game/components/PlayerGrid";
 import { AnimationSprite } from "~/components";
-import { useTimeout } from "~/styles/animations";
+import { useTimeout } from "~/utils/useTimeout";
 import { useNextStep } from "~/lib/game/useGameActions";
 import type { StepProps } from "~/lib/game/steps/types";
