+++ web/src/pages/_document.tsx
@@ -1,84 +1,54 @@
-import Document, {
-  Html,
-  Head,
-  Main,
-  NextScript,
-  DocumentContext,
-} from "next/document";
-import { ServerStyleSheet } from "styled-components";
+import { Html, Head, Main, NextScript } from "next/document";
 
-export default class MyDocument extends Document {
-  static async getInitialProps(ctx: DocumentContext) {
-    const sheet = new ServerStyleSheet();
-    const originalRenderPage = ctx.renderPage;
+export default function Document() {
+  return (
+    <Html lang="en" dir="ltr">
+      <Head>
+        <meta name="referrer" content="origin" />
+        <meta name="application-name" content="Truffles" />
+        <meta name="robots" content="index, follow" />
+        <meta property="fb:app_id" content="{FB_ID}" />
 
-    try {
-      ctx.renderPage = () =>
-        originalRenderPage({
-          enhanceApp: (App) => (props) =>
-            sheet.collectStyles(<App {...props} />),
-        });
-
-      const initialProps = await Document.getInitialProps(ctx);
-      return {
-        ...initialProps,
-        styles: [initialProps.styles, sheet.getStyleElement()],
-      };
-    } finally {
-      sheet.seal();
-    }
-  }
-
-  render() {
-    return (
-      <Html lang="en" dir="ltr">
-        <Head>
-          <meta name="referrer" content="origin" />
-          <meta name="application-name" content="Truffles" />
-          <meta name="robots" content="index, follow" />
-          <meta property="fb:app_id" content="{FB_ID}" />
-
-          <link
-            rel="apple-touch-icon"
-            sizes="180x180"
-            href="/favicon/apple-touch-icon.png"
-          />
-          <link
-            rel="icon"
-            type="image/png"
-            sizes="32x32"
-            href="/favicon/favicon-32x32.png"
-          />
-          <link
-            rel="icon"
-            type="image/png"
-            sizes="16x16"
-            href="/favicon/favicon-16x16.png"
-          />
-          <link rel="manifest" href="/favicon/site.webmanifest" />
-          <link
-            rel="mask-icon"
-            href="/favicon/safari-pinned-tab.svg"
-            color="#7247C4"
-          />
-          <link rel="shortcut icon" href="/favicon/favicon.ico" />
-          <meta name="msapplication-TileColor" content="#7247C4" />
-          <meta
-            name="msapplication-config"
-            content="/favicon/browserconfig.xml"
-          />
-          <meta name="theme-color" content="#1A202C" />
-        </Head>
-        <body className="bg-white text-black dark:bg-black dark:text-white overflow-hidden">
-          <Main />
-          <NextScript />
-          <script
-            defer
-            src="https://static.cloudflareinsights.com/beacon.min.js"
-            data-cf-beacon='{"token": "de66e4012fd9462a94e1ad0fbc4ace00"}'
-          ></script>
-        </body>
-      </Html>
-    );
-  }
+        <link
+          rel="apple-touch-icon"
+          sizes="180x180"
+          href="/favicon/apple-touch-icon.png"
+        />
+        <link
+          rel="icon"
+          type="image/png"
+          sizes="32x32"
+          href="/favicon/favicon-32x32.png"
+        />
+        <link
+          rel="icon"
+          type="image/png"
+          sizes="16x16"
+          href="/favicon/favicon-16x16.png"
+        />
+        <link rel="manifest" href="/favicon/site.webmanifest" />
+        <link
+          rel="mask-icon"
+          href="/favicon/safari-pinned-tab.svg"
+          color="#7247C4"
+        />
+        <link rel="shortcut icon" href="/favicon/favicon.ico" />
+        <meta name="msapplication-TileColor" content="#7247C4" />
+        <meta
+          name="msapplication-config"
+          content="/favicon/browserconfig.xml"
+        />
+        <meta name="theme-color" content="#1A202C" />
+      </Head>
+      <body className="bg-white text-black dark:bg-black dark:text-white overflow-hidden">
+        <Main />
+        <NextScript />
+        <script
+          defer
+          src="https://static.cloudflareinsights.com/beacon.min.js"
+          data-cf-beacon='{"token": "de66e4012fd9462a94e1ad0fbc4ace00"}'
+        ></script>
+      </body>
+    </Html>
+  )
 }
