+++ web/src/utils/useIsDesktop.ts
@@ -1,0 +1,32 @@
+import { useEffect } from "react";
+import { useDebounce } from "@react-hook/debounce";
+
+const breakpoints = {
+  desktop: "@media screen and (min-width: 900px)",
+};
+
+const isDesktop = () => {
+  if (typeof window !== "undefined" && window.matchMedia) {
+    const mq = window.matchMedia(breakpoints.desktop.replace("@media ", ""));
+    return mq.matches;
+  }
+  return false;
+};
+
+export const useIsDesktop = () => {
+  const [desktop, setDesktop] = useDebounce(isDesktop());
+
+  useEffect(() => {
+    const onResize = () => {
+      setDesktop(isDesktop());
+    };
+
+    window.addEventListener("resize", onResize);
+
+    return () => {
+      window.removeEventListener("resize", onResize);
+    };
+  }, []);
+
+  return desktop;
+};
