+++ web/src/utils/useTimeout.ts
@@ -1,0 +1,23 @@
+import { useRef, useEffect, useCallback } from "react";
+
+export const useTimeout = (
+  callback: () => void,
+  timeout: number = 0
+): (() => void) => {
+  const timeoutIdRef = useRef<NodeJS.Timeout>();
+
+  const cancel = useCallback(() => {
+    const timeoutId = timeoutIdRef.current;
+    if (timeoutId) {
+      timeoutIdRef.current = undefined;
+      clearTimeout(timeoutId);
+    }
+  }, [timeoutIdRef]);
+
+  useEffect(() => {
+    timeoutIdRef.current = setTimeout(callback, timeout);
+    return cancel;
+  }, []);
+
+  return cancel;
+};
